move handler serializers to manfiest app
