--- server/handler/serializers/__init__.py
@@ -1,2 +1,0 @@
-from .handler_serializer import HandlerSerializer, TransporterSerializer
-from .signatures import ESignatureSerializer, PaperSignatureSerializer, QuickerSignSerializer
