+++ server/manifest/serializers/__init__.py
@@ -1,1 +1,3 @@
+from .handler_serializer import HandlerSerializer, TransporterSerializer
 from .manifest import ManifestSerializer, MtnSerializer
+from .signatures import ESignatureSerializer, PaperSignatureSerializer, QuickerSignSerializer
