+++ server/manifest/tests/test_models.py
@@ -3,10 +3,10 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 from django.core.exceptions import ValidationError
-
-from handler.serializers import HandlerSerializer
-from manifest.models import Manifest, draft_mtn, manifest_factory, validate_mtn
 from rcrasite.models import RcraSiteType
+
+from manifest.models import Manifest, draft_mtn, manifest_factory, validate_mtn
+from manifest.serializers import HandlerSerializer
 
 
 @pytest.mark.django_db
